Refactor initial index
