+++ src/components/AppLayout/ListFilters/components/ExpressionFilters.tsx
@@ -1,4 +1,4 @@
-import { ConditionalFilters } from "@dashboard/components/ConditionalFilter";
+import { ConditionalProductFilters } from "@dashboard/components/ConditionalFilter";
 import { Box, Button, Popover } from "@saleor/macaw-ui/next";
 import React from "react";
 
@@ -10,7 +10,7 @@ export const ExpressionFilters = () => (
     <Popover.Content align="start">
       <Box __minWidth="200px" __minHeight="100px" paddingX={4} paddingY={3}>
         <Popover.Arrow />
-        <ConditionalFilters />
+        <ConditionalProductFilters />
       </Box>
     </Popover.Content>
   </Popover>
